start moving fuzz test cases
