+++ libz-rs-sys/src/tests/test-data/fireworks.jpg
