+++ libz-rs-sys/src/tests/test-data/lcet10.txt
